// services/react-frontend/src/library/common/types/settingsTypes.ts
export enum SettingsTypes {
	SAVE_CAMERAS = '@@SETTINGS/SAVE_CAMERAS',
	SET_ACTIVE_CAMERA_ID = '@@SETTINGS/SET_ACTIVE_CAMERA_ID',
	REQUEST_OCR_MEASUREMENT = '@@SETTINGS/REQUEST_OCR_MEASUREMENT',
	REQUEST_OCR_MEASUREMENT_SUCCESS = '@@SETTINGS/REQUEST_OCR_MEASUREMENT_SUCCESS',
	REMOVE_CALIBRATION = '@@SETTINGS/REMOVE_CALIBRATION',
	SET_IS_CALIBRATION_ACTIVE = '@@SETTINGS/SET_IS_CALIBRATION_ACTIVE',
	SET_IS_CALIBRATION_ACTIVE_SUCCESS = '@@SETTINGS/SET_IS_CALIBRATION_ACTIVE_SUCCESS',
	SET_PICTURE_LABEL = '@@SETTINGS/SET_PICTURE_LABEL',
	SET_PICTURE_LABEL_SUCCESS = '@@SETTINGS/SET_PICTURE_LABEL_SUCCESS',
	REMOVE_PICTURE_LABEL = '@@SETTINGS/REMOVE_PICTURE_LABEL',
	SET_MAGNIFICATION = '@@SETTINGS/SET_MAGNIFICATION',
}